requirements.txt
fastapi
uvicorn
python-multipart
pillow
pdf2image
pytesseract
langgraph
langchain
langchain_google_genai
python-dotenv
typing
pydantic
datetime